# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cooperative Services API",
    "description" : "Runnable prototype. Development OTP and simulated payments are explicitly gated. See docs/api-contract.md for response examples and integration rules.",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/workers/me/location" : {
      "put" : {
        "tags" : [ "worker-controller" ],
        "operationId" : "location",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Location"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/workers/me/onboarding" : {
      "post" : {
        "tags" : [ "worker-controller" ],
        "operationId" : "onboard",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Onboard"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/jobs/{id}/complete" : {
      "post" : {
        "tags" : [ "job-controller" ],
        "operationId" : "complete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/workers/me/availability" : {
      "patch" : {
        "tags" : [ "worker-controller" ],
        "operationId" : "availability",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Availability"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/me" : {
      "get" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "me",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "profile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Profile"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/workers/me" : {
      "get" : {
        "tags" : [ "worker-controller" ],
        "operationId" : "profile_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/workers/me/earnings" : {
      "get" : {
        "tags" : [ "worker-controller" ],
        "operationId" : "earnings",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/jobs/{id}/payment" : {
      "get" : {
        "tags" : [ "billing-controller" ],
        "operationId" : "payment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/catalog/subservices" : {
      "get" : {
        "tags" : [ "catalog-controller" ],
        "operationId" : "services",
        "parameters" : [ {
          "name" : "categoryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/workers" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "workers",
        "parameters" : [ {
          "name" : "societyId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Location" : {
        "type" : "object",
        "properties" : {
          "latitude" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 90,
            "minimum" : -90
          },
          "longitude" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 180,
            "minimum" : -180
          }
        },
        "required" : [ "latitude", "longitude" ]
      },
...
      "Onboard" : {
        "type" : "object",
        "properties" : {
          "societyId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "membershipId" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          },
          "uan" : {
            "type" : "string",
            "pattern" : "[0-9]{12}"
          },
          "categoryIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "maxItems" : 10,
            "minItems" : 0,
            "uniqueItems" : true
          },
          "certifications" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 200,
              "minLength" : 0
            },
            "maxItems" : 20,
            "minItems" : 0
          }
        },
        "required" : [ "categoryIds", "certifications", "membershipId", "societyId", "uan" ]
      },
...
      "Availability" : {
        "type" : "object",
        "properties" : {
          "available" : {
            "type" : "boolean"
          }
        }
      },
      "Profile" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "preferredLang" : {
            "type" : "string",
            "pattern" : "en|hi|ta"
          }
        },
        "required" : [ "name", "preferredLang" ]
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}